Re-organized the folder structure for contests page. Updated the tables on the site to be inline with the theme of the site
