--- src/Components/drawer/contests/contestsSchedule/Table.scss
@@ -1,28 +1,0 @@
-table {
-  border-collapse: collapse;
-  border-spacing: 0;
-  width: 100%;
-  border: 1px solid #ddd;
-}
-
-.rounded-corners {
-  border-radius: 8px;
-}
-
-// Making the text of each row left-aligned
-th, td {
-  text-align: left;
-  padding: 8px;
-}
-
-th {
-  color: white;
-  background-color: black;
-}
-
-tr {
-  border-bottom:1px solid #ddd;
-}
-
-//Colour of the even row
-tr:nth-child(even){background-color: #f2f2f2}
