+++ src/Components/drawer/contests/mySubmissions/index.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Cell, Grid, Row } from '@material/react-layout-grid';
 import SubmissionStatusTable from './SubmissionStatusTable';
-import ContestDetails from '../contestPage/ContestDetails';
-import Announcements from '../contestPage/Announcements';
+import ContestDetails from '../common/ContestDetails';
+import Announcements from '../common/Announcements';
 import 'tachyons';
 
 const ContestPage = (props) => {
