+++ src/Components/drawer/contests/schedule/index.jsx
@@ -1,17 +1,16 @@
 import React from 'react';
 import { Headline6, Headline4 } from '@material/react-typography';
 import { currentContestsArray, pastContestsArray } from './contestsArray';
-import Table from './Table';
-import './Table.scss';
+import ContestsTable from './ContestsTable';
 import 'tachyons';
 
 const ContestsScehdule = () => (
   <div className="pa2 mw8-ns center">
     <Headline4 className="purple mt2">Contests</Headline4>
     <Headline6 className="purple mb2 mt1">Ongoing Contests</Headline6>
-    <Table rows={currentContestsArray} />
+    <ContestsTable rows={currentContestsArray} />
     <Headline6 className="purple mb2">Past Contests</Headline6>
-    <Table rows={pastContestsArray} />
+    <ContestsTable rows={pastContestsArray} />
   </div>
 );
